requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
tabulate==0.9.0
urllib3==2.0.4
